If there is no field to preview, suggest that they add a field
Maybe if this were more effortless,
it would allow them to click a <button> to do that.
